💄 [FE] style: img 입력 윗부분으로 올리기
